requirements.txt
allure-pytest==2.9.43
pandas==1.3.0
pytest==6.2.4
requests==2.25.1
selenium==3.141.0
openpyxl==3.0.7
pytest-dependency==0.5.1
pytest-rerunfailures==10.1
cpca==0.5.5
